feat(refactoring): 구조 변경
