+++ app/api/image/image_crud.py
@@ -1,7 +1,7 @@
 import pendulum
 from dotenv import load_dotenv
 from fastapi import HTTPException
-from image.image_schema import ImageCreate
+from api.image.image_schema import ImageCreate
 from models import ContentImage, Image, User, UserImage
 from sqlalchemy.exc import SQLAlchemyError
 from sqlalchemy.orm import Session
