+++ app/models.py
@@ -1,4 +1,4 @@
-from database.database_init import Base
+from app.config.database_init import Base
 from sqlalchemy import Boolean, Column, DateTime, Integer, String
